requirements.txt
ollama
fastapi
uvicorn
supabase
python-dotenv